+++ changelog.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - improved configuration for dynamically creating CA so the directory is defaulted if not set and log output is clearer
 - improved UI handling of match failures with a because section and more complex log events
 - improved log configuration during startup when loading of properties file
+- simplified support for multiline regex by allow . to match newlines
+- improved regex matching by support Unicode (instead of US-ASCII) and native case-insensitive matching
+- improved performance of negative matches by reducing the number of regex matches when not matching
 
 ### Fixed
 - worked around JDK error 'flip()Ljava/nio/ByteBuffer; does not exist in class java.nio.ByteBuffer'